pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "platygeno"
version = "0.1.0"
description = "Unsupervised Gene Discovery via Evo 2 & SAE Interpretability"
dependencies = [
    "torch",
    "numpy",
    "biopython",
    "pandas",
    "tqdm",
    "huggingface_hub",  # Required for SAE weight auto-download
    "evo2",             # Arc Institute Evo 2 foundation model package
]
